+++ VORP-Inventory[Client-Server]/vorpinventory_sv/Diagnostics/Logger.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace vorpinventory_sv.Diagnostics
+namespace VorpInventory.Diagnostics
 {
     class Logger
     {
